third_party/WebKit/LayoutTests/inspector/console/console-correct-suggestions-expected.txt

Tests that console correctly finds suggestions in complicated cases.

Checking 'window.do'
Found: document

Checking 'win'
Found: window

Checking 'window["doc'
Found: "document"]

Checking 'window["document"].bo'
Found: body

Checking 'window["document"]["body"].textC'
Found: textContent

Checking 'document.body.inner'
Found: innerText
Found: innerHTML

Checking 'document["body"][window.do'
Found: document

Checking 'document["body"][window["document"].body.childNodes[0].text'
Found: textContent

Checking 'templateString`asdf`should'
Not Found: shouldNotFindThis

Checking 'window.document.BODY'
Found: body

Checking 'window.dOcUmE'
Found: document

Checking 'window.node'
Found: NodeList
Found: AudioNode
Found: GainNode
